Removed old viz scripts, added comparison script (#9)
This PR removes the existing visualization script, which is not working correctly anymore. As a replacement, this PR adds a comparison script to allow passing JSON files for benchmark results with two different configurations. The script transforms the JSON data into a pandas data frame, preprocesses/filters/subdivides the data frame, and generates plots.