Allow numa task nodes as matrix parmeter | Add info: allowed numa task nodes (#220)
This PR allows running mema with different numa task nodes as matrix
config parameters. Before this PR, the main thread was pinned to a
certain node after the first benchmark case. After that, in a benchmark
case where the thread is pinned to another thread,
`set_task_node_on_numa_nodes` failed since it is not in the set of nodes
the main thread is pinned to.

With this PR, we pin each "worker"/benchmark thread individually.

In addition, this PR adds info logs after the application's start
indicating what node IDs the main thread is allowed to run on.