Fix `UtilsTest.RetrieveCorrectNumaTaskNode`  (#242)
Fixes #240 and #241.

Both involved `UtilsTest.RetrieveCorrectNumaTaskNode`.
This test case should bind the test thread to all available NUMA nodes
on which threads can be pinned, one at a time.
This crashed on machines with more than one NUMA node because a check in
the second call to `numa::set_task_on_numa_nodes()` incorrectly revealed
that the test thread could not run on the requested node (because it
already has been pinned to the previous node). The solution was to reset
the NUMA run node after each iteration.
Since the test on machines with two NUMA nodes should also have failed
but didn't, it was noticed that the loop condition `node_id <
numa_max_node_id` resulted in never pinning to the node with the highest
node_id.