CI/Build: use Ruff as python linter and formatter, use poetry for dependency management (#324)
This PR introduces [Ruff](https://github.com/astral-sh/ruff) as linter
and formatter, replacing flake8 and black.
It also adds a CI action to check the formatting of the python scripts.

Ruff can be configured using ruff.toml. As before we use the default
behavior but change linelength to 120 and ignore E203.

A pyproject.toml is added to specify the python environment. All tools
can be configured from within this file.
The PR also introduces Python-Poetry for package management. Poetry
manages a lock file to keep environments in sync.